backend/requirements.txt
openai
python-dotenv
requests
flask
flask-cors